rca-pipeline/requirements.txt
# AI-Powered RCA Pipeline — Python dependencies
# Install with: pip install -r requirements.txt

# LLM client
openai==1.30.0

# Environment variable loading
python-dotenv==1.0.1

# Web framework (for Python inventory service)
flask==3.0.3

# WSGI application server (for production deployment)
gunicorn==22.0.0

# HTTP requests (used by some agents in production mode)
requests==2.31.0

# Numerical computing (used by analysis agents)
numpy==1.24.4
